short call pruning
This version only considers focal calls . Although detected calls that are shorter than the shortest calls in the ground truth fro that call category are supposedly removed, this changes almost nothing to the precision recall curves, meaning something is wrong in that regard. I'm hoping to be able to fix this in the next few days.